simplification of WTSeriesTraining
