+++ mlcf/aitools/super_module.py
@@ -13,7 +13,7 @@ from time import time_ns
 
 # MLCF modules
 from mlcf.datatools.wtseries_tensor import WTSeriesTensor
-from mlcf.datatools.wtseries_training import Partition, WTSeriesTraining
+from mlcf.datatools.wtst import Partition, WTSTraining
 from mlcf.aitools.log import ProgressBar, add_metrics_to_log, log_to_message
 from mlcf.envtools.hometools import ProjectHome
 from mlcf.aitools.training_manager import TrainingManager
@@ -131,7 +131,7 @@ class SuperModule(Module):
         self.initialize = True
 
     def fit(self,
-            dataset: WTSeriesTraining,
+            dataset: WTSTraining,
             n_epochs: int,
             batchsize: int,
             shuffle: bool = False,
@@ -143,7 +143,7 @@ class SuperModule(Module):
         """Fit/train the model (need to be initialized first)
 
         Args:
-            dataset (WTSeriesTraining): The dataset used to train (and evaluate the model)
+            dataset (WTSTraining): The dataset used to train (and evaluate the model)
             n_epochs (int): The number of epochs
             batchsize (int): The batch size
             shuffle (bool, optional): If we want to shuffle the data. Defaults to True.
@@ -166,15 +166,24 @@ class SuperModule(Module):
         if not self.initialize:
             raise Exception("The module has not been compiled")
 
-        train_data = WTSeriesTensor(Partition.TRAIN, ts_data=dataset,
-                                    transform_x=self.transform_x,
-                                    transform_y=self.transform_y)
-        validation_data = WTSeriesTensor(Partition.VALIDATION, ts_data=dataset,
-                                         transform_x=self.transform_x,
-                                         transform_y=self.transform_y)
-        test_data = WTSeriesTensor(Partition.TEST, ts_data=dataset,
-                                   transform_x=self.transform_x,
-                                   transform_y=self.transform_y)
+        train_data = WTSeriesTensor(
+            ts_data=dataset,
+            partition=Partition.TRAIN,
+            transform_x=self.transform_x,
+            transform_y=self.transform_y
+        )
+        validation_data = WTSeriesTensor(
+            ts_data=dataset,
+            partition=Partition.VALIDATION,
+            transform_x=self.transform_x,
+            transform_y=self.transform_y
+        )
+        test_data = WTSeriesTensor(
+            ts_data=dataset,
+            partition=Partition.TEST,
+            transform_x=self.transform_x,
+            transform_y=self.transform_y
+        )
 
         self.manager.debug(
                 f"\nSize train tensor (x,y): {train_data.x_size()}, {train_data.y_size()}" +
