+++ mlcf/datatools/wtseries_tensor.py
@@ -1,5 +1,4 @@
 import numpy as np
-from pathlib import Path
 from typing import Callable
 
 from torch.utils.data import TensorDataset
@@ -7,14 +6,13 @@ from torch.utils.data.dataloader import DataLoader
 from torch import Tensor, tensor
 
 # MLCF modules
-from mlcf.datatools.wtseries_training import Partition, Field, WTSeriesTraining
+from mlcf.datatools.wtst import Partition, WTSTraining
 
 
 class WTSeriesTensor(TensorDataset):
     def __init__(self,
+                 ts_data: WTSTraining,
                  partition: Partition,
-                 ts_data: WTSeriesTraining = None,
-                 dir_ts_data: Path = None,
                  transform_x: Callable = None,
                  transform_y: Callable = None,
                  *args, **kwargs):
@@ -23,11 +21,11 @@ class WTSeriesTensor(TensorDataset):
 
         Args:
             partition (str): the part 'train', 'validation' or 'test'
-            ts_data (WTSeriesTraining, optional): 'The time series data'. Defaults to None.
+            ts_data (WTSTraining, optional): 'The time series data'. Defaults to None.
             dir_ts_data (Path, optional): 'the dir to the time series file'. Defaults to None.
 
         Raises:
-            ValueError: You should to have a WTSeriesTraining (ts_data) or at least the dir to a
+            ValueError: You should to have a WTSTraining (ts_data) or at least the dir to a
             TimeSeries (dir_ts_data)
             NotImplementedError: read a time serie data from a file is not implemented yet
             Warning: read a time serie data from a file is not implemented yet
@@ -37,20 +35,13 @@ class WTSeriesTensor(TensorDataset):
         self.partition = partition
         self.transform_x = transform_x
         self.transform_y = transform_y
+        self.ts_data = ts_data
+        self.ts_data.set_partition(self.partition)
 
-        if ts_data is None and dir_ts_data is None:
-            raise ValueError("You should to have a WTSeriesTraining (ts_data) or " +
-                             "at least the dir to a TimeSeries (dir_ts_data).")
-        if dir_ts_data is not None and ts_data is None:
-            raise NotImplementedError("read a time serie data from a file is not implemented yet")
-        if ts_data is not None:
-            if dir_ts_data is not None:
-                raise Warning("read a time serie data from a file is not implemented yet")
-            self.ts_data: WTSeriesTraining = ts_data
         if len(self.ts_data) == 0:
-            raise ValueError("WTSeriesTraining has a length of 0. It is empty")
+            raise ValueError("WTSTraining has a length of 0. It is empty")
 
-        self.ts_data_to_tensor(self.partition, transform_x=transform_x, transform_y=transform_y)
+        self.ts_data_to_tensor(transform_x=transform_x, transform_y=transform_y)
 
         super(WTSeriesTensor, self).__init__(*[self.input_data, self.target_data])
 
@@ -65,14 +56,14 @@ class WTSeriesTensor(TensorDataset):
             return len(self.input_data)
         return self.ts_data.len(part=self.partition)
 
-    def ts_data_to_tensor(self,
-                          partition: Partition,
-                          transform_x: Callable = None,
-                          transform_y: Callable = None):
-        self.input_data = tensor(np.array(self.ts_data(partition,
-                                                       Field.INPUT)).astype(np.float32))
-        self.target_data = tensor(np.array(self.ts_data(partition,
-                                                        Field.TARGET)).astype(np.float32))
+    def ts_data_to_tensor(
+        self,
+        transform_x: Callable = None,
+        transform_y: Callable = None
+    ):
+        inputs, targets = self.ts_data()
+        self.input_data = tensor(np.array(inputs).astype(np.float32))
+        self.target_data = tensor(np.array(targets).astype(np.float32))
         self.transform_data(transform_x, transform_y)
 
     def get_dataloader(self, *args, **kwargs):
@@ -87,8 +78,9 @@ class WTSeriesTensor(TensorDataset):
             self.target_data = transform_y(self.target_data)
 
     def copy(self):
-        return WTSeriesTensor(self.partition,
-                              ts_data=self.ts_data,
-                              dir_ts_data=None,
-                              transform_x=self.transform_x,
-                              transform_y=self.transform_y)
+        return WTSeriesTensor(
+            ts_data=self.ts_data,
+            partition=self.partition,
+            transform_x=self.transform_x,
+            transform_y=self.transform_y
+        )
