+++ tests/datatools/test_wtseries_tensor.py
@@ -1,6 +1,6 @@
 import pytest
 import pandas as pd
-from mlcf.datatools.wtseries_training import Partition, Field, WTSeriesTraining
+from mlcf.datatools.wtst import Partition, WTSTraining
 from mlcf.datatools.wtseries_tensor import WTSeriesTensor
 import numpy as np
 
@@ -17,19 +17,20 @@ data = init_data()
 
 
 def test_WTSeriesTensor():
-    ts_data = WTSeriesTraining(9, index_column="date")
+    ts_data = WTSTraining(9, index_column="date")
     with pytest.raises(ValueError):
-        dataset = WTSeriesTensor(Partition.TRAIN, ts_data=ts_data)
+        dataset = WTSeriesTensor(ts_data=ts_data, partition=Partition.TRAIN)
     ts_data.add_time_serie(data.iloc[0:3000])
     assert ts_data.index_column == "date"
-    dataset = WTSeriesTensor(Partition.TRAIN, ts_data=ts_data)
+    dataset = WTSeriesTensor(ts_data=ts_data, partition=Partition.TRAIN)
     assert ts_data.len(Partition.TRAIN) == 2391
     assert ts_data.len(Partition.VALIDATION) == 171
     assert ts_data.len(Partition.TEST) == 411
     for i in range(len(dataset)):
         inp, tar = dataset[i]
-        assert np.all(np.array(inp) == np.array(ts_data(Partition.TRAIN,
-                                                        Field.INPUT)[i].astype(np.float32)))
-        assert np.all(np.array(tar) == np.array(ts_data(Partition.TRAIN,
-                                                        Field.TARGET)[i].astype(np.float32)))
-    assert dataset.input_data.size() == ts_data(Partition.TRAIN, Field.INPUT).size()
+        inputs, targets = ts_data(Partition.TRAIN)
+        assert np.all(np.array(inp) == np.array(inputs[i].astype(np.float32)))
+        assert np.all(np.array(tar) == np.array(targets[i].astype(np.float32)))
+    inputs, targets = ts_data(Partition.TRAIN)
+    assert dataset.input_data.size() == inputs.size()
+    assert dataset.target_data.size() == targets.size()
